ci: Switch everything over to temp envs + uv run
Switches all of the tools that run locally on our runners to rely on
temp environments, this should enable us to be able to run on any
machine that has uv and also has the benefit of not intermixing virtual
envs between different utilities we have built.

A nice follow-up to this would be to just write python utils for all of
these and include the dependencies in the front matter but I think
that's a more involved PR that I will save for when I have more time.

This adds the setup-uv action which should be idempotent to every
composite action that may need it in order to run.

I've confirmed that setup-uv should be a noop if uv is already installed
so ideally this should be a noop after its run once.

https://github.com/astral-sh/setup-uv/blob/main/src/setup-uv.ts

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: d6a9ff109ad832af15145c93d62e40eed51f4f6e
Pull-Request: https://github.com/pytorch/pytorch/pull/170316
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>