Change requires_grad to Optional[bool] in torch.asarray
The current default behavior when invoking `torch.asarray` is to drop
any existing gradients on the input. This is confusing when inputting a
tensor object, so this commits changes the default to follow the input
iff the input is a tensor (while preserving explicit override behavior).
This is technically an API breaking change, but a minor one that I
suspect will not affect downstream users.

Fixes #170819.


ghstack-source-id: 8cd89ea0b877dfe8761f7d5d3ee97b47596af650
Pull-Request: https://github.com/pytorch/pytorch/pull/170897